Found the sum and average hours played by category of experience
I culled the players dataset of the variable hashedEmail, and then used group_by and summarize to find the mean and total hours played by each category of experience.

I then used group_by and summarize to find the total amount of a subgroup in the whole dataset to ensure that the results aren't being skewed by a group overtaking in magnitude.