Made A bar graph of the relationship between type of player and average hours played
In this commit, I made a bar graph of the relationship between the level of experience a player has and their average hours played. I used average hours played instead of total amount of hours played because I believe the data yielded from that will be too skewed towards the group that takes up the majority of the dataset.